third_party/WebKit/LayoutTests/inspector-protocol/css/cssom-matching-rules-expected.txt

The test verifies CSS.getMatchedStylesForNode when used concurrently with the CSSOM modifications.


Running test: testModifyRule
--------------
Original rule:
*#modifyRule* {    regular
    box-sizing: border-box;
}
*#modifyRule* {    regular
    height: 100%;
}
*#modifyRule* {    regular
    width: 100%;
}
--------------
Modified rule 1:
*#modifyrule* {    regular    readonly
    box-sizing: border-box;
    color: red;
}
*#modifyRule* {    regular
    height: 100%;
}
*#modifyRule* {    regular
    width: 100%;
}
---------------
Modified rule 3:
*#modifyrule* {    regular    readonly
    box-sizing: border-box;
    color: red;
}
*#modifyRule* {    regular
    height: 100%;
}
*#modifyrule* {    regular    readonly
    width: 100%;
    color: blue;
}
---------------
Modified rule 2:
*#modifyrule* {    regular    readonly
    box-sizing: border-box;
    color: red;
}
*#modifyrule* {    regular    readonly
    height: 100%;
    color: green;
}
*#modifyrule* {    regular    readonly
    width: 100%;
    color: blue;
}
---------------
Restored rule 2:
*#modifyrule* {    regular    readonly
    box-sizing: border-box;
    color: red;
}
*#modifyRule* {    regular
    height: 100%;
}
*#modifyrule* {    regular    readonly
    width: 100%;
    color: blue;
}
-----------------
Restored rule 1,3:
*#modifyRule* {    regular
    box-sizing: border-box;
}
*#modifyRule* {    regular
    height: 100%;
}
*#modifyRule* {    regular
    width: 100%;
}

Running test: testInsertFirstRule
Original rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}
--------------
After inserted rule:
*#insertrule* {    regular    readonly
    color: red;
}
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}
--------------
Restored rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}

Running test: testInsertMiddleRule
Original rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}
--------------
After inserted rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertrule* {    regular    readonly
    color: red;
}
*#insertRule* {    regular
    width: 100%;
}
--------------
Restored rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}

Running test: testInsertLastRule
Original rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}
--------------
After inserted rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}
*#insertrule* {    regular    readonly
    color: red;
}
--------------
Restored rule:
*#insertRule* {    regular
    box-sizing: border-box;
}
*#insertRule* {    regular
    width: 100%;
}

Running test: testRemoveRule
Original rule:
*#removeRule* {    regular
    box-sizing: border-box;
}
*#removeRule* {    regular
    width: 100%;
}
-------------------
After remove rule 1:
*#removeRule* {    regular
    width: 100%;
}
-------------------
After remove rule 2:
